+++ src/lib/models/ReflectionCategory.ts
@@ -76,7 +76,7 @@ export class ReflectionCategory {
                     const child = project.getReflectionById(
                         de.oldIdToNewId[childId] ?? -1,
                     );
-                    if (child?.isDeclaration()) {
+                    if (child?.isDeclaration() || child?.isDocument()) {
                         this.children.push(child);
                     }
                 }