+++ CHANGELOG.md
@@ -19,6 +19,8 @@ title: Changelog
 -   Fix restoration of groups/categories including documents, #2801.
 -   Fixed missed relative paths within markdown link references in documents.
 -   Improved handling of incomplete inline code blocks within markdown.
+-   Direct `https://` links under the `hostedBaseUrl` option's URL will no
+    longer be treated as external, #2809.
 
 ### Thanks!